requirements.txt
discord.py
tweepy
flask